+++ cs-ilp-core/Org/Interledger/Encoding/Asn/Codecs/Suppliers/AsnUint32CodecSupplier.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+using Org.Interledger.Encoding.Asn.Framework;
+
+namespace Org.Interledger.Encoding.Asn.Codecs.Suppliers
+{
+    public class AsnUint32CodecSupplier : IAsnObjectCodecSupplier<uint>
+    {
+        public IAsnObjectCodec<uint> Get()
+        {
+            return new AsnUint32Codec();
+        }
+    }
+}
