+++ cs-ilp-core/Org/Interledger/Encoding/Asn/Codecs/Suppliers/AsnUint64CodecSupplier.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+using Org.Interledger.Encoding.Asn.Framework;
+
+namespace Org.Interledger.Encoding.Asn.Codecs.Suppliers
+{
+    public class AsnUint64CodecSupplier : IAsnObjectCodecSupplier<ulong>
+    {
+        public IAsnObjectCodec<ulong> Get()
+        {
+            return new AsnUint64Codec();
+        }
+    }
+}
