+++ cs-ilp-core/cs-ilp-core.csproj
@@ -17,6 +17,7 @@
     <Folder Include="Org\Interledger\Encoding\Asn\Serializers\" />
     <Folder Include="Org\Interledger\Encoding\Asn\Serializers\Oer\" />
     <Folder Include="Org\Interledger\Core\PSK2\" />
+    <Folder Include="Org\Interledger\Encoding\Asn\Codecs\Suppliers\" />
   </ItemGroup>
   <ItemGroup>
     <Compile Remove="Org\Interledger\Core\SharedSecret.cs" />
